Move metric related code to metrics.utils.py
