+++ segmentation/utils.py
@@ -2,7 +2,6 @@ import commentjson as json
 from importlib import import_module
 import numpy as np
 import torch
-import torch.nn.functional as F
 import logging
 import sys
 import os
@@ -138,56 +137,6 @@ def save_checkpoint(state, save_path, model):
         model.save(filename)
     else:
         torch.save(state, filename)
-
-
-def channel_metrics(prediction, target, metric):
-    """
-    Compute a given metric on every channel of the volumes.
-    """
-    if target.shape[1] == 1:
-        target = torch.cat([target, 1 - target], dim=1)
-    prediction = F.softmax(prediction, dim=1)
-    channels = list(range(target.shape[1]))
-    res = []
-    for channel in channels:
-        res.append(metric(prediction[:, channel, ...], target[:, channel, ...]))
-
-    return torch.tensor(res)
-
-
-def between_channel_metrics(prediction, target, metric, cut=0.5):
-    """
-    Compute a given metric on every channel using masks computed from the other channels.
-    """
-    if target.shape[1] == 1:
-        target = torch.cat([target, 1 - target], dim=1)
-    prediction = F.softmax(prediction, dim=1)
-    channels = list(range(target.shape[1]))
-    res = []
-    for i in channels:
-        for j in channels:
-            if i != j:
-                mask = target[:, j, ...] > cut
-            else:
-                mask = (target[:, j, ...] > 0) * (target[:, j, ...] < 1)
-            res.append(metric(prediction[:, i, ...] * mask, target[:, i, ...] * mask))
-
-    return torch.tensor(res)
-
-
-def mean_metric(prediction, target, metric):
-    """
-    Compute a given metric on every channel of the volumes and average them.
-    """
-    if target.shape[1] == 1:
-        target = torch.cat([target, 1 - target], dim=1)
-    prediction = F.softmax(prediction, dim=1)
-    channels = list(range(target.shape[1]))
-    res = 0
-    for channel in channels:
-        res += metric(prediction[:, channel, ...], target[:, channel, ...])
-
-    return res / len(channels)
 
 
 def get_class_name_from_method(method):
